moved release_info to checkbox support
